Add verbosity option

+++ lib/image_renamer.rb
@@ -7,17 +7,14 @@ class ImageRenamer
     @images = images
   end
 
-  def chronological!(dry_run: false)
+  def chronological!(dry_run: false, verbose: false)
     @images.
       map { |image| [image, ExifDateReader.call(image)] }.
       sort_by { |file, date| date }.
       each_with_index do |(file, date), index|
         newname = date.strftime(filename_template(index+1))
-        if dry_run
-          puts "Renaming #{file} to #{newname}"
-        else
-          FileUtils.mv(file, File.join(File.dirname(file), newname))
-        end
+        puts "#{file} -> #{newname}" if verbose
+        FileUtils.mv(file, File.join(File.dirname(file), newname)) unless dry_run
       end
   end
